backend/requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
pywebpush
cryptography
passlib[bcrypt]
